+++ host/class/uvc/usb_host_uvc/uvc_descriptor_parsing.c
@@ -130,9 +130,9 @@ int uvc_desc_parse_format(const uvc_format_desc_t *format_desc)
 {
     // Input checks
     assert(format_desc);
-    UVC_CHECK(uvc_desc_is_format_desc((const usb_standard_desc_t *)format_desc), UVC_VS_FORMAT_UNDEFINED);
+    UVC_CHECK(uvc_desc_is_format_desc((const usb_standard_desc_t *)format_desc), -1);
 
-    int ret = UVC_VS_FORMAT_UNDEFINED;
+    int ret = -1;
     switch (format_desc->bDescriptorSubType) {
     case UVC_VS_DESC_SUBTYPE_FORMAT_UNCOMPRESSED: {
         const char *guid = (const char *)(format_desc->uncompressed_frame_based.guidFormat);
@@ -156,8 +156,6 @@ int uvc_desc_parse_format(const uvc_format_desc_t *format_desc)
         break;
     }
     case UVC_VS_DESC_SUBTYPE_UNDEFINED:
-        ret =  UVC_VS_FORMAT_UNDEFINED;
-        break;
     default: break;
     }
     return ret;
@@ -169,6 +167,11 @@ static bool uvc_desc_format_is_equal(const uvc_frame_desc_t *frame_desc, const u
     UVC_CHECK(uvc_desc_is_frame_desc((const usb_standard_desc_t *)frame_desc), false);
 
     if (frame_desc->wWidth == vs_format->h_res && frame_desc->wHeight == vs_format->v_res) {
+        if (vs_format->fps == 0) {
+            return true; // Default FPS requested
+        }
+
+        // Check if this frame descriptor supports requested FPS
         uint8_t bFrameIntervalType;
         uint32_t dwMinFrameInterval, dwMaxFrameInterval, dwFrameIntervalStep;
 
@@ -223,8 +226,7 @@ static bool uvc_desc_format_is_equal(const uvc_frame_desc_t *frame_desc, const u
             }
             break;
         default:
-            assert(!"This subtype is not supported!");
-            return false;
+            return true; // Unknown frame type, assume it is equal;
         }
     }
     return false;
@@ -277,7 +279,11 @@ static inline bool uvc_desc_is_format_supported(
     uint8_t bInterfaceNumber,
     const uvc_host_stream_format_t *vs_format)
 {
-    return (ESP_OK == uvc_desc_get_frame_format_by_format(cfg_desc, bInterfaceNumber, vs_format, NULL, NULL));
+    if (vs_format->format == UVC_VS_FORMAT_DEFAULT ||
+            ESP_OK == uvc_desc_get_frame_format_by_format(cfg_desc, bInterfaceNumber, vs_format, NULL, NULL)) {
+        return true;
+    }
+    return false;
 }
 
 static const uvc_vc_header_desc_t *uvc_desc_get_control_interface_header(const usb_config_desc_t *cfg_desc, unsigned uvc_idx)
@@ -420,7 +426,7 @@ esp_err_t uvc_desc_get_frame_list(const usb_config_desc_t *config_desc, uint8_t 
 
         const uvc_format_desc_t *this_format = (const uvc_format_desc_t *)(current_desc);
         enum uvc_host_stream_format format_type = uvc_desc_parse_format(this_format);
-        if (UVC_VS_FORMAT_UNDEFINED == format_type) {
+        if (0 > format_type) { // Undefined format
             continue;
         }